Make arm_dyncom_trans* into a fully fledged compilation unit
